Remove unneeded files.
Thumbs.db is a cache generated by Windows.

The tmp/ folder is automatically generated by RenPy and isn't needed to build the game itself.

traceback.txt is a file generated by RenPy that contains the last error encountered while playing the game. This is mainly used to help debug your game, but isn't used to build the game itself.